// frontend/src/app/api/model/pagedResponseReviewViewDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ReviewViewDto } from './reviewViewDto';


export interface PagedResponseReviewViewDto { 
    content?: Array<ReviewViewDto>;
    pageNumber?: number;
    pageSize?: number;
    totalElements?: number;
    totalPages?: number;
    isLast?: boolean;
}